Switch to Manual Synchronization
Previously, the system used Continuous Syncing, which is limited to 200 bytes per serialization. I switched to Manual Syncing, as it expands the data limit from 200 -> 280,496 bytes. For scale, a character is supposedly 2 bytes per character at minimum, plus added bytes to the total for networking due to Serialization Overhead. So if "Mass Joining! Please be careful, these could be trolls! Come tell me if they are!" is sent, ~174–184+ bytes were sent over networking, so a longer message previously wouldn't sync do to clogging. Additionally, the Announcement System now waits for data to be properly sent before attempting to display notifications to the local player. This update makes the Announcement System overall more reliable.