Bug Fix 1.01
Attempted to patch a bug. If a player leaves the instance, it is off by 1 player. For example, before leaving, it displays 7 players, but after someone leaves, it would still report 7 players when there are 6 players now. Added a small delay before checking and updating the counter. Should've just subtracted one; if this isn't patched, I'll do that instead.